refactor/87: rename SelectedCounter component
